Topic/rev/feature/integrate crud on addresses pickups (#10)
This story is about CRUD on pickups, addresses and profile information of a User. It integrates the PickUp and addresses tables with the UI design in order to conveniently do the viewing of pickups and address (both default and non default) and modification of addresses (both default and non default).