Integrate Auth framework for the UI (#12)
This story is about integration of authentication framework wrt all the features like booking, my cases, invoices and logout.
This auth token is stored locally and tracks checking of authentication on every event which requests server. In case of session expiry it takes the user to logout and clears the auth token there by no more calls are authenticated.