Create New Case and upload STL files and create prescription. Make dynamic setting of proxy for Axios. (#16)
This story is about creating new case for an orthodontist and able to upload STL files for a case in the font end.As the proxy is for dev purposes, we have inculcated the dynamic setting of proxy for the staging purposes.

After retrieving orthodontist's information, we add all the information in the case Image in HTML and convert it to an image with the text overlay over it. Using Axios we upload STL files and the generated prescription image to the Google cloud platform in the backend.